Back to simpler lock files

+++ sync.sh
@@ -8,20 +8,28 @@ time=`date +%H:%M:%S`
 logfile="$basedir/logs/$location-sync-$date.log"
 
 # Don't run more than one sync script at a time
-if [ -z "$flock" ] ; then
-	lockopts="-w 0 $lockfile"
-	exec env flock=1 flock $lockopts $0 "$@"
+if [ -f "$lockfile" ] ; then
+	echo "Lock file exists: $lockfile"
+	exit 1
 fi
+
+touch "$lockfile"
 
 {
 	echo "-- $date $time --"
 
 	# Sync video files
-	rsync -rvi --ignore-existing "$basedir/videos/$location" flserver:/home/flightlines/videos/
+	rsync -rv --ignore-existing "$basedir/videos/$location" flserver:/home/flightlines/videos/
 
 	# Update scripts
-	cd $basedir && git pull origin master -q
-} >> $logfile
+	cd "$basedir" && git pull origin master -q
+} >> "$logfile"
 
 # Sync log files
 rsync -r --exclude .keep-dir "$basedir/logs/" "flserver:/home/flightlines/$location/"
+
+date=`date +%Y-%m-%d`
+time=`date +%H:%M:%S`
+echo "Finished at $date $time"
+
+rm "$lockfile"
